requirements.txt
langchain-core
langchain-community
langchain-openai
langchain-aws
streamlit
python-dotenv==1.0.0
boto3
sqlalchemy==2.0.21
psycopg2-binary
aws-cdk-lib>=2.0.0
constructs